% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.documentation.r
\docType{data}
\name{biological.oceanographic.data}
\alias{biological.oceanographic.data}
\title{biological.oceanographic.data}
\format{A list with several elements}
\usage{
biological.oceanographic.data$Riki_Zoo_Mean_2018
}
\description{
A list with biological oceanographic time series objects for the AE ecoregions of the Gulf of St. Lawrence.
The list contains several data frame with descriptive names and named columns. An example of how to access
and one of the data objects is shown. Missing value code is na
}
\details{
\itemize{
  \item Riki_Zoo_Mean_2018 annual mean abundance of zooplankton species or species group abundance at the
       Rimouski station updated in 2018.
}
}
\author{
Marjolaine Blais <marjolaine.blais@dfo-mpo.gc.ca>
}
\keyword{datasets}
