% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.r
\name{cold.over.warm.f}
\alias{cold.over.warm.f}
\title{Cold to warm water species biomass ratio}
\usage{
cold.over.warm.f()
}
\description{
The following species were identified as candidate indicator taxa for cold water habitat in
      the EGSL: Boreogadus saida, Artediellus uncinatus, Eumicrotremus terraenovae, Pandalus borealis and
      Reinhardtius hippoglossoides. The following species were identified as candidate indicator taxa for
      warm water habitat in the EGSL: Merluccius bilinearis, Illex illecebrosus, Argentina silus.

      The biomass of each candidate indicator taxa in each trawl-year combination was calculated. These
      can be examined as individual indicator taxa or as indicator groups (by summing the biomass of all
      cold or warm indicator taxa for each trawl-year combination).

      Next steps: Calculate frequency of occurrence and index of relative importance (IRI) for these taxa
      at different spatial scales. Calculate and examine trajectories of mean and max length for candidate
      indicator taxa. Produce and investigate the usefulness of a length-unbiased condition factor for
      candidate indicator taxa.
}
