% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.documentation.r
\docType{data}
\name{EA.data}
\alias{EA.data}
\title{Ecosystem Approach Data}
\format{data.table}
\usage{
EA.data
}
\description{
A data.table with all the ecosystem.approach data for the Gulf of St. Lawrence ecoregions. This includes data on
physical oceanography (physical), chemical oceanographic (chemical), phyto and zooplankton abundance or biomass
(planktonic) and timing of phyto and zooplankton dynamics (phenological).
}
\details{
\itemize{
  \item year The year of data collection
  \item EAR The Ecosystem Approach Region
  \item variable The name of the measured variable
  \item value The value of the measured variable
}
}
\seealso{
field.description
}
\author{
Peter Galbraith, Marjolaine Blais
}
\keyword{approach}
\keyword{data}
\keyword{ecosystem}
