% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{EAR.name.location.f}
\alias{EAR.name.location.f}
\title{Plot data time series}
\usage{
EAR.name.location.f()
}
\description{
Joins the English names of the ecoregions to the EA.data to aid graph descriptions. It also joins
              columns for  the longitude (- degrees from prime meridian) and latitude (+ degrees from equator)
              of the centre of each ecoregion polygon.
}
\examples{
EA.data2= EAR.name.location.f()
}
\author{
Daniel Duplisea
}
