% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{find.vars.f}
\alias{find.vars.f}
\title{Find variables with a keyword or phrase}
\usage{
find.vars.f(search.term)
}
\arguments{
\item{search.term}{a term to search, e.g. "oxygen", "temperature", "temp", "saturat"}
}
\description{
Does a fuzzy search for the word in the variable description and returns the full variable names.
}
\examples{
find.vars.f("temp")
       find.vars.f("satu")
}
\author{
Daniel Duplisea
}
