% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.documentation.r
\docType{data}
\name{fish.survey.data}
\alias{fish.survey.data}
\title{fish.survey.data}
\format{A list with several elements}
\description{
A list with fish survey data. Imported from PACES outputs.
}
\details{
\itemize{
  \item r
}
}
\author{
Denis Bernier <denis.bernier@dfo-mpo.gc.ca>
}
\keyword{datasets}
