% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{gslea.map.f}
\alias{gslea.map.f}
\title{a map of the Gulf of St. Lawrence with ecoregions and options to include survey points and bathymetry}
\usage{
gslea.map.f(
  xlim = c(-70, -55),
  ylim = c(42, 53),
  bathymetry = F,
  isobath = c(150, 300),
  bathycols = c("yellow", "blue")
)
}
\arguments{
\item{bathymetry}{show bathymetry lines}

\item{isobath}{bamthymetric depths to show}

\item{bathycols}{a vector of bathymetric colour, recycled}

\item{survey.points}{logical. Plot the fish survey points from all years from the set file}

\item{pch}{the point type for depicting fish survey points}
}
\description{
A map of the Gulf of St Lawrence showing the ecoregions and with options to plot different
       bathymetric contours and fish survey points
}
