% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.documentation.r
\name{gslea}
\alias{gslea}
\title{Gulf of St Lawrence Ecosystem Approach R package}
\usage{
gslea()
}
\description{
To get started you probably want some kind of metadata description:

                  metadata.f(verbosity="low")


             You may be interested in what variables are available:

                  vars.f(variable.type="all")

             You may be interested in what variables are available of a particular type:

                  vars.f(variable.type="chemical")


             You may want to explore data with simple plots:

                  EA.plot.f(years=1900:2030, variables=c("T150", "ph_bot.fall", "T250"), EARs=1:4, type="b",pch=20)


             You may just want to extract the data from the above plot:

                  EA.query.f(variables=c("T150", "ph_bot.fall", "T250"), years=1900:2030, EARs=1:4)
}
