% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.r
\name{guild.biomass.f}
\alias{guild.biomass.f}
\title{Guild biomass}
\usage{
guild.biomass.f()
}
\description{
The biomass of each functional guild was calculated for each trawl-year combination. Biomass
      trajectories for individual functional guilds, or ratios of functional guilds biomass to total
      biomass in each trawl-year, may be used as ecological indicators (e.g., the proportion of predatory
      fish in individual trawl (bio_pel.bpel.pisc+bio_dem.bent.pisc /Trawl.Biomass)
}
\details{
Fish and invertebrate species were classified into 11 functional guilds:
     Planktivorous fish (F) (bio_pkt.fish)
     Benthivorous fish (F) (bio_benthiv.fish)
     Pelagic and Benthopelagic piscivores (F, I) (bio_pel.bpel.pisc)
     Demersal and Benthic piscivores (F, I) (bio_dem.bent.pisc)
     Parasitic and Detritivorous fish (F) (bio_par.det)
     Benthic infauna (I) (bio_benthic.infauna)
     Benthic motile invertebrates (I) (bio_benthic.mot.inv)
     Crab (I) (bio_crab)
     Demersal and benthic Micronekton (I) (bio_dem.bent.micronekton)
     Pelagic and benthopelagic Micronekton (I) (bio_pel.bpel.micronekton)
     Sessile invertebrates (I) (bio_sess.inv)
}
\references{
Baker, M. R., and A. B. Hollowed. 2014. Delineating ecological regions in marine systems:
      Integrating physical structure and community composition to inform spatial management in the eastern
      Bering Sea. Deep Sea Research Part II: Topical Studies in Oceanography 109:215-240.

      Coll, M., L. Shannon, K. Kleisner, M. Juan-Jordá, A. Bundy, A. Akoglu, D. Banaru, J. Boldt,
      M. Borges, and A. Cook. 2016. Ecological indicators to capture the effects of fishing on biodiversity
      and conservation status of marine ecosystems. Ecological Indicators 60:947-962.
}
