% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.r
\name{inv.cv.f}
\alias{inv.cv.f}
\title{Inverse biomass temporal variance}
\usage{
inv.cv.f()
}
\description{
This could be an interesting one to compute when scaling up to AE boxes and the entire
     ecosystem.
}
\references{
Coll, M., L. Shannon, K. Kleisner, M. Juan-Jordá, A. Bundy, A. Akoglu, D. Banaru, J. Boldt,
      M. Borges, and A. Cook. 2016. Ecological indicators to capture the effects of fishing on biodiversity
      and conservation status of marine ecosystems. Ecological Indicators 60:947-962.
}
