% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.r
\name{inv.over.fish.f}
\alias{inv.over.fish.f}
\title{Invertebrate to fish biomass ratio}
\usage{
inv.over.fish.f()
}
\description{
The ratio of total invertebrates biomass to total fish biomass in each trawl. Individual trawl
      biomass was calculated and standardized in kg per km2, based on individual tow swept area (product of
      tow distance derived from tow duration and trawling speed and trawl horizontal opening or wing spread).
}
