% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{metadata.f}
\alias{metadata.f}
\title{Metadata with three levels of verbosity}
\usage{
metadata.f(verbosity = "low")
}
\arguments{
\item{verbosity}{"low","med","high".}
}
\description{
Pulls out metadata from the various datasets to give an overview. "med" and "high" verbosity level data
              may need further display formatting to get a good overview. Try the library 'formattable' and run that
              function on the higher verbosity metadata to get a nicer looking table that you can make sense of.
}
\examples{
metadata.f("low")
data.desc= metadata.f("med"); library(formattable); formattable(data.desc$Descriptions, align="l")
metadata.f("vangogh")
}
\author{
Daniel Duplisea
}
