% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.r
\name{pel.over.dem.f}
\alias{pel.over.dem.f}
\title{pelagic to demersal fish biomass ratio}
\usage{
pel.over.dem.f()
}
\description{
Individual fish species were classified into four vertical habitat groupings: 1) benthic
      species living and feeding on the bottom; 2) demersal species living and feeding on or near the
      bottom; 3) benthopelagic species living and feeding near the bottom as well as in midwaters or near
      the surface; and 4) pelagic species living and feeding in midwaters or near the surface.
}
