% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.documentation.r
\docType{data}
\name{physical.oceanographic.data}
\alias{physical.oceanographic.data}
\title{physical.oceanographic.data}
\format{A list with several elements}
\usage{
physical.oceanographic.data$CIL_Volume_AE$data
}
\description{
A list with physical oceanographic time series objects for the AE ecoregion of the Gulf of St. Lawrence.
The list contains several sublists with descriptive names. The first element of each of these sublists is
a descriptive header file including what it is, date of extraction, who to contact and column names. The
second element of each sublist is a data frame with the time series data and descriptive (but very long)
column names. An example of how to access and one of the data objects is shown. Missing value code is -99
}
\details{
\itemize{
  \item CIL_Volume_AE Volume of the Cold Intermediate Layer
}
}
\author{
Peter Galbraith <peter.galbraith@dfo-mpo.gc.ca>
}
\keyword{datasets}
