% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{sources.f}
\alias{sources.f}
\title{Sources and references for each variable}
\usage{
sources.f(variable.name = NULL)
}
\description{
creates a three column data.table with variable, source and reference columns. Please cite the authors if you use their data.
}
\examples{
sources.f()
}
\author{
Daniel Duplisea
}
