% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.r
\name{species.diversity.f}
\alias{species.diversity.f}
\title{Species richness and diversity}
\usage{
species.diversity.f()
}
\description{
This approach reduced the total number of individual taxa recorded in the trawl survey data
      in 2006-2018 from 477 to 252. Richness was calculated as the total number of individual taxa in each
      trawl-year combination. This index could be further improved by i) distinguishing commercial
      invertebrate taxa from the broader taxonomic groups described above; and ii) by further refining
      and standardising invertebrate taxonomic resolution, for example by identifying dominant taxa that
      should be distinguished to species versus infrequent taxa that could remain pooled into higher-order
      taxonomic groups.
      Continuous improvement and standardisation of taxonomic resolution in the trawl survey data will also
      allow for the development of reliable, fisheries-independent species diversity indicators for the
      EGSL.
}
\details{
For the purpose of evaluating species richness, invertebrate taxa belonging to the following three
      functional guilds were pooled into higher-order taxonomic groupings in order to minimise the effect
      of recently increasing taxonomic resolution for these guilds in the survey data:
      Sessile invertebrates were pooled into the following 11 groups: Actiniaria, Alyconacea, Ascidiacea,
      Bryozoa, Cirripedia, Crinoidea, Hydrozoa, Pennatulacea, Porifera, Scleractinia and Zoanthidea.

      Benthic motile invertebrates were pooled into the following 9 groups: Asteroidea, Cumacea,
      Echinoidea, Gastropoda, Holothuroidea, Hydrozoa, Ophiuroidea, Picnogonida and Polyplacophora.

      Benthic infauna were pooled into the following 10 groups: Bivalvia, Brachiopoda, Echiura, Isopoda,
      Nematoda, Nemertea, Polycheate, Priapulidae, Scaphopoda and Sipuncula.
}
