% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.r
\name{TL2.overTL3.f}
\alias{TL2.overTL3.f}
\title{Trophic level 2 to Trophic levels 3 and more fish biomass ratio}
\usage{
TL2.overTL3.f()
}
\description{
Fish species were classified into four feeding categories: 1) Planktivorous; 2) Benthivorous;
      3) Piscivorous one and two (where piscivorous-one refers to shrimp (micronekton) consumers and
      piscivorous-two refers to both shrimp and fish consumers); and 4) Detritivorous/Scavenger. Species
      that undergo ontogenetic dietary shifts were assigned to a feeding category based on their adult
      stage diet (e.g., Sebastes sp. switch from a planktivorous to a piscivorous diet with increasing
      size and were assigned to the piscivorous category).
}
