% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.documentation.r
\docType{data}
\name{variable.description}
\alias{variable.description}
\title{Variable descriptions in EA.data}
\format{data.table, data.frame}
\usage{
variable.description
}
\description{
Descriptions of the variables in EA.data, the units of measure and other information
}
\details{
\itemize{
  \item variable The name of the variable in EA.data
  \item description A description of the variable
  \item units The units of measure for the value of the variable
  \item contact The name of the contact person for each variable
  \item type The broad category of the variabl, e.g. "physical", "chemical", "planktonic", "phenological"
  \item extraction.date The date which the data were extracted from the parent data base by the person named as the contact
}
}
\seealso{
field.description
}
\author{
Daniel Duplisea
}
\keyword{approach}
\keyword{data}
\keyword{ecosystem}
