% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{vars.f}
\alias{vars.f}
\title{Show  variables available in the data set}
\usage{
vars.f(variable.type = "all")
}
\arguments{
\item{type}{The type of data you are interested in. Options: "all","physical","chemical","planktonic","phenological","fish"}
}
\description{
Lists all the variables in the database by data type
}
\examples{
vars.f("phenological")
#consider using the library formattable to make a readable table that prints to the plot window
tmp= vars.f("all"); formattable(tmp)
}
\author{
Daniel Duplisea
}
