% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.documentation.r
\docType{data}
\name{EA.data}
\alias{EA.data}
\title{Ecosystem Approach Data}
\format{
data.table, data.frame
}
\usage{
EA.data
}
\description{
A data.table with all the ecosystem.approach data for the Gulf of St. Lawrence ecoregions. This includes data on
physical oceanography (physical); chemical oceanographic (chemical); phyto and zooplankton abundance or biomass
(planktonic); timing of phyto and zooplankton dynamics (phenological); large scale indices of oceanography, atmosphere
at broad scales (climatic); atmospheric projections from the climateatlas.ca website that brings together predictions from
24 global climate models.
}
\details{
\itemize{
  \item year The year of data collection
  \item EAR The Ecosystem Approach Region
  \item variable The name of the measured variable
  \item value The value of the measured variable
}
}
\seealso{
field.description
}
\keyword{approach}
\keyword{data}
\keyword{ecosystem}
