% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.documentation.r
\docType{data}
\name{EA.data}
\alias{EA.data}
\title{Ecosystem Approach Data}
\format{data.table, data.frame}
\usage{
EA.data
}
\description{
A data.table with all the ecosystem.approach data for the Gulf of St. Lawrence ecoregions. This includes data on
physical oceanography (physical), chemical oceanographic (chemical), phyto and zooplankton abundance or biomass
(planktonic) and timing of phyto and zooplankton dynamics (phenological).
}
\details{
\itemize{
  \item year The year of data collection
  \item EAR The Ecosystem Approach Region
  \item variable The name of the measured variable
  \item value The value of the measured variable
}
}
\references{
Blais, M., Galbraith, P.S., Plourde, S., Scarratt, M., Devine, L. and Lehoux, C. 2019. Chemical and
        Biological Oceanographic Conditions in the Estuary and Gulf of St. Lawrence during 2017. DFO Can. Sci.
        Advis. Sec. Res. Doc. 2019/009. iv + 56 pp.

        Galbraith, P.S., Chassé, J., Nicot, P., Caverhill, C., Gilbert, D., Pettigrew, B., Lefaivre,
        D., Brickman, D., Devine, L., and Lafleur, C. 2015. Physical Oceanographic Conditions in the Gulf of
        St. Lawrence in 2014. DFO Can. Sci. Advis. Sec. Res. Doc. 2015/032. v + 82 p
        http://www.dfo-mpo.gc.ca/csas-sccs/Publications/ResDocs-DocRech/2015/2015_032-eng.html
}
\seealso{
field.description
}
\author{
Denis Bernier, Peter Galbraith, Marjolaine Blais
}
\keyword{approach}
\keyword{data}
\keyword{ecosystem}
