% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.documentation.r
\docType{data}
\name{EA.data}
\alias{EA.data}
\title{Ecosystem Approach Data}
\format{
\itemize{
  \item year The year of data collection
  \item EAR The Ecosystem Approach Region
  \item variable The name of the measured variable
  \item value The value of the measured variable
}
}
\usage{
EA.data
}
\description{
A data.table with data useful in developing analysese that could be classified as an ecosystem approach data for the Gulf of St. Lawrence ecoregions.
}
\details{
This includes data on
physical oceanography (physical); chemical oceanographic (chemical); phyto and zooplankton abundance or biomass
(planktonic); large scale indices of oceanography, atmosphere
at broad scales (climatic); atmospheric projections from the climateatlas.ca website that brings together predictions from
24 global climate models (projection.atmospheric); fish survey biomass estimates are divided into three groups: core species
(fish.survey.core.species) are 19 species that are consistently captured by the survey, commercial species (fish.survey.commercial.species)
includes 9 species which have been or are commercially exploited. The commercial species biomass is provided for juveniles and adults with the division between juvenile
and adult being the regulations on the small fish protocol:

Cod/ Morue – 43 cm
American Plaice/Plie canadienne – 30 cm
Witch Flounder/ Plie grise – 30 cm
White hake / Merluche blanche – 45 cm
Winter Flounder / Plie rouge – 25 cm
Yellowtail Flounder/ Limande à queue jaune - 25 cm
Atlantic Halibut/ Flétan de l’Atlantique – 85 cm
Greenland Halibut/ Flétan du Groenland – 44 cm
Redfish/ Sébaste – 22 cm.

Finally fish biomass is also aggregated into functional guilds (fish.survey.guild) which is a hybrid guild concept based on
both the main food source of species in that guild as well as their primary habitat.
}
\seealso{
field.description
}
\keyword{approach}
\keyword{data}
\keyword{ecosystem}
