% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{EA.plot.f}
\alias{EA.plot.f}
\title{Plot data time series}
\usage{
EA.plot.f(variables, years, EARs, smoothing = T, ...)
}
\arguments{
\item{variables}{variable vector}

\item{years}{year vector}

\item{EARs}{EAR vector}

\item{smoothing}{if TRUE and n>5, then a smooth.spline is drawn through the data with df=n/3}

\item{...}{arguments to par for plotting}
}
\description{
Plots a matrix of variables with EAR as columns and variable as rows. If no data then a blank graph
              is plotted.
}
\examples{
EA.plot.f(variables=c("T150", "ph_bot.fall", "T250"), years=1900:2030, EARs=1:4, type="b",pch=20)
}
\author{
Daniel Duplisea
}
