% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{EA.query.f}
\alias{EA.query.f}
\title{Query the data}
\usage{
EA.query.f(variables, years, EARs)
}
\arguments{
\item{variables}{variable vector}

\item{years}{year vector}

\item{EARs}{EAR vector}
}
\description{
creates a data.table based on the variables you selected. data.tables inherit data.frame
}
\examples{
EA.query.f(years=1999:2012, variables=c("T150", "ph_bot.fall", "T250"), EARs=1:2)
}
\author{
Daniel Duplisea
}
