% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{find.vars.f}
\alias{find.vars.f}
\title{Find variables with a keyword or phrase}
\usage{
find.vars.f(search.term, description = FALSE)
}
\arguments{
\item{search.term}{a term to search, e.g. "oxygen", "temperature", "temp", "saturat". can have multiple terms, e.g. c("coD","250","MaRjO")}

\item{description}{if T then the description of the variable is also provided (this can be long). Default FALSE.}

\item{units}{if T then the units of measure of the variable is also provided. Default FALSE.}
}
\description{
Does a search for the string(s) in the variable description and returns the full variable names.
}
\examples{
find.vars.f("temp", description=T)
       find.vars.f("satu")
}
\seealso{
like, %like%, grepl, variable.description
}
\author{
Daniel Duplisea
}
